pyproject.toml
[project]
name = "fabriquinha-de-certificados"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.1",
    "fastapi>=0.115.9",
    "mangum>=0.19.0",
    "psycopg>=3.2.5",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.1",
    "qrcode[pil]>=8.0",
    "sqlalchemy>=2.0.39",
    "weasyprint>=64.1",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.7",
    "types-qrcode>=8.0.0.20241004",
]

[tool.pytest.ini_options]
addopts = "-vvv --cov-report=term-missing --cov=fabriquinha"
testpaths = ["tests"]
log_level = "DEBUG"
